infra-cost-monitor/slack_bot/requirements.txt
# Slack Bot Dependencies
slack-bolt==1.18.1
slack-sdk==3.26.1

# RAG System Integration
sentence-transformers==2.2.2
chromadb==0.4.22

# LLM Integration
ollama==0.5.1

# Data Processing
pandas==2.0.3
numpy==1.24.4

# Web Framework
flask==3.0.3
flask-cors==5.0.0

# Utilities
python-dotenv==1.0.1
requests==2.32.4
python-dateutil==2.8.2

# Development
pytest==7.4.3 